engine_traits: Rename CFHandleExt to CFOptionsExt
