+++ components/engine_panic/src/lib.rs
@@ -11,8 +11,6 @@
 
 #![allow(unused)]
 
-mod cf_handle;
-pub use crate::cf_handle::*;
 mod cf_names;
 pub use crate::cf_names::*;
 mod cf_options;
