+++ components/engine_rocks/src/lib.rs
@@ -27,8 +27,6 @@ extern crate serde_derive;
 #[cfg(test)]
 extern crate test;
 
-mod cf_handle;
-pub use crate::cf_handle::*;
 mod cf_names;
 pub use crate::cf_names::*;
 mod cf_options;
