+++ components/engine_traits/src/cf_options.rs
@@ -1,6 +1,15 @@
 // Copyright 2019 TiKV Project Authors. Licensed under Apache-2.0.
 
 use crate::{db_options::TitanDBOptions, sst_partitioner::SstPartitionerFactory};
+use crate::Result;
+
+/// Trait for engines with column family options
+pub trait CFOptionsExt {
+    type ColumnFamilyOptions: ColumnFamilyOptions;
+
+    fn get_options_cf(&self, cf: &str) -> Result<Self::ColumnFamilyOptions>;
+    fn set_options_cf(&self, cf: &str, options: &[(&str, &str)]) -> Result<()>;
+}
 
 pub trait ColumnFamilyOptions {
     type TitanDBOptions: TitanDBOptions;
@@ -12,7 +21,7 @@ pub trait ColumnFamilyOptions {
     fn get_soft_pending_compaction_bytes_limit(&self) -> u64;
     fn get_hard_pending_compaction_bytes_limit(&self) -> u64;
     fn get_block_cache_capacity(&self) -> u64;
-    fn set_block_cache_capacity(&self, capacity: u64) -> Result<(), String>;
+    fn set_block_cache_capacity(&self, capacity: u64) -> std::result::Result<(), String>;
     fn set_titandb_options(&mut self, opts: &Self::TitanDBOptions);
     fn get_target_file_size_base(&self) -> u64;
     fn set_disable_auto_compactions(&mut self, v: bool);
